feat: Closes #24
