+++ Cargo.toml
@@ -16,6 +16,8 @@ tabled = "0.15"
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 ambassador = "0.3.6"
+serde_json = "1.0.115"
+serde = "1.0.197"
 
 [features]
 default = ["cli"]
