+++ src/managers/mod.rs
@@ -56,9 +56,9 @@ impl MetaPackageManager {
             AvailablePackageManager::Zypper => Self::Zypper(Zypper),
         };
 
-        match mpm.cmd().arg("--version").status() {
-            Ok(status) => {
-                if status.success() {
+        match mpm.cmd().arg("--version").output() {
+            Ok(output) => {
+                if output.status.success() {
                     Ok(mpm)
                 } else {
                     anyhow::bail!("failed to run {mpm} command")
